+++ apps/web/src/components/core/TextArea/TextArea.tsx
@@ -10,7 +10,7 @@ import {
 import styled from 'styled-components';
 
 import colors from '~/shared/theme/colors';
-import noop from '~/utils/noop';
+import { noop } from '~/shared/utils/noop';
 
 import MarkdownShortcuts, { setValueAndTriggerOnChange } from '../Markdown/MarkdownShortcuts';
 import { BaseM } from '../Text/Text';
