--- apps/web/src/utils/noop.ts
@@ -1,3 +1,0 @@
-export default function noop() {
-  // No operation
-}
