+++ packages/shared/src/utils/noop.ts
@@ -1,0 +1,3 @@
+export function noop() {
+  // No operation
+}
